pyproject.toml
[project]
name = "claude-voice-hooks"
version = "0.3.0"
description = "Claude Code hook system with audio feedback and TTS integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core dependencies
    "accelerate==1.12.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "audioop-lts==0.2.2",
    "audioread==3.1.0",
    "av==16.0.1",
    "brotli==1.2.0",
    "budoux==0.7.0",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cn2an==0.5.23",
    "colorama==0.4.6",
    "coloredlogs==15.0.1",
    "colorlog==6.10.1",
    "contourpy==1.3.3",
    "ctranslate2==4.6.2",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "distance==0.1.3",
    "einops==0.8.1",
    "einx==0.3.0",
    "fast-langdetect==0.2.5",
    "fastapi==0.124.4",
    "faster-whisper==1.2.1",
    "fasttext-predict==0.9.2.4",
    "ffmpeg-python==0.2.0",
    "ffmpy==1.0.0",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "fonttools==4.61.1",
    "frozendict==2.4.7",
    "frozenlist==1.8.0",
    "fsspec==2025.12.0",
    "future==1.0.0",
    "g2p-en==2.1.0",
    "gradio==6.1.0",
    "gradio-client==2.0.1",
    "groovy==0.1.2",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-resources==6.5.2",
    "inflect==7.5.0",
    "jieba==0.42.1",
    "jieba-fast==0.53",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "librosa==0.11.0",
    "lightning-utilities==0.15.2",
    "llvmlite==0.46.0",
    "loguru==0.7.3",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "matplotlib==3.10.8",
    "mdurl==0.1.2",
    "more-itertools==10.8.0",
    "mpmath==1.3.0",
    "msgpack==1.1.2",
    "multidict==6.7.0",
    "networkx==3.5",
    "nltk==3.9.2",
    "numba==0.63.1",
    "numpy==2.3.5",
    "onnxruntime==1.23.2",
    "orjson==3.11.5",
    "packaging==25.0",
    "pandas==2.3.3",
    "peft==0.18.0",
    "pillow==12.0.0",
    "platformdirs==4.5.1",
    "pooch==1.8.2",
    "proces==0.1.7",
    "propcache==0.4.1",
    "protobuf==6.33.2",
    "psutil==7.1.3",
    "pycparser==2.23",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "pyopenjtalk==0.4.1",
    "pyparsing==3.2.5",
    "pypinyin==0.55.0",
    "pyreadline3==3.5.4",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "pytorch-lightning==2.6.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "rich==14.2.0",
    "robust-downloader==0.0.2",
    "safehttpx==0.1.7",
    "safetensors==0.7.0",
    "scikit-learn==1.8.0",
    "scipy==1.16.3",
    "semantic-version==2.10.0",
    "setuptools==70.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "soundfile==0.13.1",
    "soxr==1.0.0",
    "split-lang==2.0.6",
    "standard-aifc==3.13.0",
    "standard-chunk==3.13.0",
    "standard-sunau==3.13.0",
    "starlette==0.50.0",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "tomlkit==0.13.3",
    "torch==2.7.1+cu118",
    "torchaudio==2.7.1+cu118",
    "torchmetrics==1.8.2",
    "tqdm==4.67.1",
    "transformers==4.57.3",
    "typeguard==4.4.4",
    "typer==0.20.0",
    "typer-slim==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.6.2",
    "uvicorn==0.38.0",
    "win32-setctime==1.2.0",
    "x-transformers==2.11.23",
    "yarl==1.22.0",
    # Additional dependencies for translation API
    "ollama>=0.4.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# PyTorch requires custom index for CUDA builds
index-url = "https://download.pytorch.org/whl/cu118"
extra-index-url = ["https://pypi.org/simple"]

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
